clientapi into multiple controllers
